% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-helpers.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Return data as data.frame}
\usage{
as_data_frame(d)
}
\arguments{
\item{d}{data to be converted to data frame}
}
\description{
Return data as data.frame
}
\keyword{internal}
