% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_plot.R
\name{extract_y_axis}
\alias{extract_y_axis}
\title{Extract pretty y-axis information from d}
\usage{
extract_y_axis(meta)
}
\arguments{
\item{d}{data}
}
\description{
Extract pretty y-axis information from d
}
\keyword{internal}
