% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-helpers.R
\name{get_param}
\alias{get_param}
\title{Get appropriate parameter value}
\usage{
get_param(name, lst, default)
}
\arguments{
\item{name}{name of parameter}

\item{lst}{list with parameter/value tuples}

\item{default}{default value that is returned if 'name' not present in 'lst'}
}
\value{
parameter from 'lst' or if non-present 'default'
}
\description{
Get appropriate parameter value
}
