% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-helpers.R
\name{is_valid_extension}
\alias{is_valid_extension}
\title{Check if the xlsx file is valid}
\usage{
is_valid_extension(file_name)
}
\arguments{
\item{file_name}{xlsx file name you want to check}
}
\value{
Boolean indicating whether xlsx is valid
}
\description{
Check if the xlsx file is valid
}
\keyword{internal}
