% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_activate.R
\name{j_activate}
\alias{j_activate}
\title{Activate a certain project and scenario}
\usage{
j_activate(type = james.env$j_root$type,
  scenario = james.env$j_root$scenario,
  project = james.env$j_root$project)
}
\arguments{
\item{type}{the data type you want to activate}

\item{scenario}{the scenario you want to activate}

\item{project}{the project you want to activate}
}
\value{
None
}
\description{
Activate a certain project and scenario
}
