% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_example_xlsx.R
\name{j_example_xlsx}
\alias{j_example_xlsx}
\title{Create example xlsx file}
\usage{
j_example_xlsx(file_name = paste0(tempfile(), ".xlsx"),
  multiple_tabs = TRUE)
}
\arguments{
\item{file_name}{where your xlsx-file should be stored (default temporary xlsx-file)}

\item{multiple_tabs}{creates an xlsx-file with multiple tabs if TRUE (default), including a meta tab. If FALSE, it creates an xlsx-file with only one tab}
}
\value{
Path to your example xlsx
}
\description{
You can use this example to import, analyse, and visualise your own data
}
\seealso{
\code{\link{j_import}} to import the example data
}
