% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_fig.R
\name{j_fig}
\alias{j_fig}
\title{Easy plotting in your standard layout}
\usage{
j_fig(index)
}
\arguments{
\item{index}{id of data you want to plot}
}
\value{
None
}
\description{
Plot the data at position 'index' in j_ls(), given its meta data
}
