% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_get_index.R
\name{j_get_index}
\alias{j_get_index}
\title{Get index}
\usage{
j_get_index(type, version, scenario = james.env$j_root$active_scenario,
  project = james.env$j_root$active_project)
}
\arguments{
\item{type}{data type (default "")}

\item{version}{version of data you want index of (default most recent version only)}

\item{scenario}{scenario filter (defaults to active scenario)}

\item{project}{project filter (defaults to active project)}
}
\value{
index
}
\description{
Get index of data in j_ls()
}
