% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_get_ts.R
\name{j_get_ts}
\alias{j_get_ts}
\title{Get data as ts object}
\usage{
j_get_ts(index, type, version, scenario = james.env$j_root$scenario,
  project = james.env$j_root$project)
}
\arguments{
\item{index}{unique id of your data in j_ls()}

\item{type}{data type you want to get}

\item{version}{version you want to get (default is active (c.q. newest) version)}

\item{scenario}{scenario from which you want to get the data (default is active scenario)}

\item{project}{project from which you want to get the data (default is active project)}
}
\value{
index
}
\description{
Same as j_get() but with result converted to ts object
}
