% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_import.R
\name{j_import}
\alias{j_import}
\title{Import data from xlsx-file}
\usage{
j_import(file_name, meta)
}
\arguments{
\item{file_name}{xlsx-file you want to import}

\item{meta}{meta parameters used to import, e.g. meta = list(project = "A project", scenario = "Test"). Beware these meta parameters overwrite imported meta parameters if they share the same name.}
}
\value{
None
}
\description{
j_import()
}
