% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_import_settings.R
\name{j_import_settings}
\alias{j_import_settings}
\title{Import settings}
\usage{
j_import_settings(file_name)
}
\arguments{
\item{file_name}{xlsx-file you want to import}
}
\value{
Indices of imported data in j_ls()
}
\description{
Import settings
}
