% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_ls.R
\name{j_ls}
\alias{j_ls}
\title{List James' data}
\usage{
j_ls(type, version, scenario, project, collapsed = TRUE,
  active_project_scenario_only = TRUE)
}
\arguments{
\item{type}{data type}

\item{version}{version of data you want to filter on}

\item{scenario}{scenario filter (defaults to active scenario), see also active_project_scenario_only}

\item{project}{project filter (defaults to active project), see also active_project_scenario_only}

\item{collapsed}{shows only most recent version of each data type (default TRUE)}

\item{active_project_scenario_only}{if project/scenario are missing, list only data in your active project and active scenario (default TRUE)}
}
\value{
data.frame which describes the data in james.env$j_root, given the parameters
}
\description{
Equivalent of R's native ls() function.
}
\seealso{
\code{\link{j_init}} to initialise James. See \code{\link{j_put}} and \code{\link{j_get}} to get/put data
}
