% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_plot.R
\name{j_plot}
\alias{j_plot}
\title{Easy plotting in your standard layout}
\usage{
j_plot(index, meta)
}
\arguments{
\item{index}{id of data you want to plot}

\item{meta}{settings you want to use for your figure. These settings are not stored and will be used only once.}
}
\value{
None
}
\description{
Plot the data at position 'index' in j_ls(), given its meta data
}
