% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_put.R
\name{j_put}
\alias{j_put}
\title{Put data in archive}
\usage{
j_put(x, type = james.env$j_root$type,
  scenario = james.env$j_root$scenario,
  project = james.env$j_root$project, doc = NA,
  add_if_duplicate = TRUE)
}
\arguments{
\item{x}{data you want to store}

\item{type}{(choose something useful, e.g. "input")}

\item{scenario}{the scenario (default active scenario)}

\item{project}{project to which x belongs (default active project)}

\item{doc}{document your data with plain text or Markdown}

\item{add_if_duplicate}{store if data are already present? Prevents storing duplicates, e.g. after re-running your script (default TRUE)}
}
\value{
index of (added) data. Both invisibly.
}
\description{
Put data in archive at specified location. TODO: add all meta data with j_put
}
\seealso{
\code{\link{j_get}}, \code{\link{j_ls}}, \code{\link{j_init}}
}
