% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_save.R
\name{j_save}
\alias{j_save}
\title{Save data to file}
\usage{
j_save()
}
\value{
None
}
\description{
Save data you have put in archive to file that was specified in j_init()
}
\examples{
example(j_get)

}
\seealso{
\code{\link{j_init}} to initialise James. See \code{\link{j_put}} and \code{\link{j_get}} to get/put data
}
