% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_set_meta.R
\name{j_set_meta}
\alias{j_set_meta}
\title{Set meta data}
\usage{
j_set_meta(index, fields)
}
\arguments{
\item{index}{unique id of your data in j_ls()}

\item{fields}{list with fields you want to add or edit}
}
\value{
None
}
\description{
Set meta data for figure
}
