% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_snapshot.R
\name{j_snapshot}
\alias{j_snapshot}
\title{Export xlsx-data file}
\usage{
j_snapshot(is_final = FALSE, create_figs = TRUE,
  include_figs = FALSE, open_data_file = FALSE, meta = list(),
  input_file_name = "cepmev.xlsx", snapshot_dir = file.path("snapshot",
  format(Sys.time(), "\%Y\%m\%d-\%H\%M")),
  data_file_name = guess_doc_specs()$file_name,
  doc_name = guess_doc_specs()$doc_name,
  publication_date = guess_doc_specs()$publication_date)
}
\arguments{
\item{is_final}{(Boolean) prepends user (read: dirnames) if FALSE}
}
\value{
data_file_name
}
\description{
Optionally with png's inside.
}
