% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_validate.R
\name{j_validate}
\alias{j_validate}
\title{Validate meta data}
\usage{
j_validate(index)
}
\arguments{
\item{index}{unique id of your data in j_ls()}
}
\value{
error message or NULL if none
}
\description{
Validate meta data
}
