% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_plot.R
\name{extract_x_axis}
\alias{extract_x_axis}
\title{Extract x-axis information from d}
\usage{
extract_x_axis(d)
}
\arguments{
\item{d}{retrieve x-axis}
}
\description{
Extract x-axis information from d
}
\keyword{internal}
