% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_activate.R
\name{j_activate}
\alias{j_activate}
\title{Activate a certain project and scenario}
\usage{
j_activate(type = james.env$j_root$type,
  scenario = james.env$j_root$scenario, project = james.env$j_root$project)
}
\arguments{
\item{type}{the data type you want to activate}

\item{scenario}{the scenario you want to activate}

\item{project}{the project you want to activate}
}
\value{
None
}
\description{
j_ls() lists only data from the active project and scenario. If you change project and/or scenario, you can use this function to activate it/them. Use unique(j_ls(active_project_scenario_only=F)[c("project", "scenario")]) to list all unique combinations of the two.
}
