% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_example_xlsx.R
\name{j_example_xlsx}
\alias{j_example_xlsx}
\title{Get example xlsx file}
\usage{
j_example_xlsx(file_name = paste0(tempfile(), ".xlsx"))
}
\arguments{
\item{file_name}{file_name where your xlsx-file should be stored (default temporary xlsx-file)}
}
\value{
None
}
\description{
You can use this example to import, analyse, and visualise your own data
}
