% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_fig.R
\name{j_fig}
\alias{j_fig}
\title{Easy plotting in your standard layout}
\usage{
j_fig(index, format = c(NA, "pdf"), file_name)
}
\arguments{
\item{index}{id of data you want to plot}

\item{format}{format in which you want to save your figure. Default is NA, meaning plot to screen}

\item{file_name}{file name}
}
\value{
None
}
\description{
Plot the data at position 'index' in j_ls(), given its meta data
}
