% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_import_settings.R
\name{j_import_settings}
\alias{j_import_settings}
\title{Import settings}
\usage{
j_import_settings(meta = list())
}
\arguments{
\item{meta}{list with meta data}
}
\value{
list with meta data
}
\description{
Import settings from xlsx-file and add them to meta.
}
\details{
\code{meta} gets highest prio. Settings in \code{file_name} get lower prio. If you don't specify \code{file_name}, james will use the james-settings.xlsx from its own package.
}
