% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_import_settings.R
\name{j_import_settings}
\alias{j_import_settings}
\title{Import settings}
\usage{
j_import_settings(file_name, meta = list())
}
\arguments{
\item{file_name}{xlsx-file with settings you want to import (if not specified, a file with defaults in package will be imported)}

\item{meta}{list with meta data}
}
\value{
list with meta data
}
\description{
Import settings from xlsx-file and add them to meta.
}
\details{
\code{meta} gets highest prio. Settings in \code{file_name} get lower prio. If you don't specify \code{file_name}, James will use james-settings.xlsx (in current directory). Side effect: if this file is non-existent, James will put a copy of its default version there.
}
