% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_ls.R
\name{j_ls}
\alias{j_ls}
\title{List James' data}
\usage{
j_ls(type, version, scenario, project, collapse = FALSE,
  filter_active = FALSE)
}
\arguments{
\item{type}{data type (default "")}

\item{version}{version of data you want to filter on}

\item{scenario}{scenario filter (defaults to active scenario if filter_active = TRUE)}

\item{project}{project filter (defaults to active project if filter_active = TRUE)}

\item{collapse}{show only most recent version of each data type (default FALSE)}

\item{filter_active}{if project/scenario are missing, list only data in your active project and active scenario (default FALSE)}
}
\value{
data.frame which describes the data in james.env$j_root, given the parameters
}
\description{
Equivalent of R's native ls() function.
}
\seealso{
\code{\link{j_activate}} to activate another project/scenario
}
