% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_put.R
\name{j_put}
\alias{j_put}
\title{Put data in archive}
\usage{
j_put(x, type = "", doc = NA, scenario = james.env$j_root$active_scenario,
  project = james.env$j_root$active_project, add_if_duplicate = TRUE,
  activate_project_scenario = TRUE)
}
\arguments{
\item{x}{data you want to store}

\item{type}{(choose something useful, e.g. "input")}

\item{doc}{document your data with plain text or Markdown}

\item{scenario}{the scenario (default active scenario)}

\item{project}{project to which x belongs (default active project)}

\item{add_if_duplicate}{store if data are already present? Prevents storing duplicates, e.g. after re-running your script (default TRUE)}

\item{activate_project_scenario}{the given project and scenario become active after storing x (default TRUE), which sets defaults for the next time you want to get/put data}
}
\value{
index if successfully added, else NULL. Both invisibly.
}
\description{
Put data in archive at specified location. Beware
}
\seealso{
\code{\link{j_get}}, \code{\link{j_ls}}, \code{\link{j_init}}
}
