% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_plot.R
\name{one_element_for_each_series}
\alias{one_element_for_each_series}
\title{Meta may have variables of length 1 or length > n_items. This function makes all lists have length n_items.}
\usage{
one_element_for_each_series(meta)
}
\description{
Meta may have variables of length 1 or length > n_items. This function makes all lists have length n_items.
}
\keyword{internal}
